chore : 페이지네이션 관련 유틸 함수 한 파일로 모으기
